fix(product): keep the launch-intent view model out of the login first-load graph (PRO-230)
The launch-intent registry is reached from the signed-out shell's eager graph
(ProductLifecycleRoot -> use-app-shortcuts -> workspace selection -> session
creation -> chat-launch-intent-store -> launch-intent-registry), and it needed
two functions out of launch-intent.ts. Importing them dragged that whole module
— retry-mode resolution, the rendered view model, the catalog snapshot types —
into the login entry chunk, which has a fail-closed JS budget.

Shell ownership moves to launch-intent-scope.ts and the registry imports it
directly; launch-intent.ts re-exports both so no other importer changes. That is
transcript-time work no login screen can reach, and it now stays out of the
first-load chunk.
